UPDATED DEVLOG
Update on missing days and up-to-date:

Day 2 & 3 (15/07 & 16/07)
Worked on a side Unity project to test out an Enemy AI idea I had to make the enemy be smart. The idea sadly did not work but the overall concept did. Enemies detected the player, went near the player, and kept within a circle radius around the player and would try to attack without going beyond the radius. The radius itself caused a problem where enemies ended up orbiting around the player and not attacking it properly - attacks either just didn't hit the player, or if the attack range was increased, would behave weirdly and phase through the entire radius and onto the other side.
I attempted to fix this for both Day 2, 3 and parts of Day 4(17/07) before deciding to take a break for the remainder of Day 4 so that I could work on the original idea (Enemy Behaviour Trees.) with a clear head.

Day 5 (18/07)
Began work initially during the late evening of Day 4 on the Behaviour Tree. I have the base code for a behaviour tree set up and just need to add more 'leafs' (Actions.) to it before implementing it fully.

Development has come to a small standstill due to adding a NavMesh Package (NavMeshPlus) to allow for enemy pathfinding.